requirements.txt
discord.py==2.3.2
fortnitepy==3.6.6
motor==3.3.2
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
aiofiles==23.2.1
